refactor(developer): move functions and variables into CompilerErrors.cpp
Groundwork for rework of compiler errors. No functional changes.
