+++ common/windows/delphi/keyboards/kmxfileconsts.pas
@@ -29,7 +29,6 @@ const
   SZMAX_GROUPNAME = 80;
   SZMAX_DEADKEYNAME = 80;
   SZMAX_VKDICTIONARYNAME = 80;  // I3438
-  SZMAX_ERRORTEXT = 512;
 
   SZMAX_LANGUAGENAME = 80;
   SZMAX_KEYBOARDNAME = 80;
