+++ developer/src/kmcmplib/src/Compiler.cpp
@@ -110,18 +110,13 @@
 #define PRIMARYLANGID(lgid)    ((uint16_t)(lgid) & 0x3ff)
 #define SUBLANGID(lgid)        ((uint16_t)(lgid) >> 10)
 
-#define COMPILE_ERROR_MAX_LEN (SZMAX_ERRORTEXT + 1 + 280)
-
 using namespace kmcmp;
 
-  char ErrExtraLIB[ERR_EXTRA_LIB_LEN]; // utf-8
   KMX_BOOL AWarnDeprecatedCode_GLOBAL_LIB;
 
 namespace kmcmp{
   KMX_BOOL  FShouldAddCompilerVersion = TRUE;
   KMX_BOOL  FSaveDebug, FCompilerWarningsAsErrors;   // I4865   // I4866
-  int ErrChr;
-  int nErrors = 0;
   KMX_BOOL FMnemonicLayout = FALSE;
   KMX_BOOL FOldCharPosMatching = FALSE;
   int CompileTarget;
@@ -270,49 +265,6 @@ PKMX_STR wstrtostr(PKMX_WCHAR in)
   return result;
 }
 
-KMX_BOOL kmcmp::AddCompileWarning(PKMX_CHAR buf)
-{
-  (*msgproc)(kmcmp::currentLine + 1, CWARN_Info, buf, msgprocContext);
-  return FALSE;
-}
-
-KMX_BOOL AddCompileError(KMX_DWORD msg)
-{
-  KMX_CHAR szText[COMPILE_ERROR_MAX_LEN];
-  const KMX_CHAR* szTextp = NULL;
-
-  if (msg & CERR_FATAL)
-  {
-    szTextp = GetCompilerErrorString(msg);
-    (*msgproc)(kmcmp::currentLine + 1, msg, szTextp, msgprocContext);
-    kmcmp::nErrors++;
-    return TRUE;
-  }
-
-  if (msg & CERR_ERROR)
-    kmcmp::nErrors++;
-  szTextp = GetCompilerErrorString(msg);
-
-  if (szTextp) {
-    strcpy(szText, szTextp);
-  } else {
-    snprintf(szText, COMPILE_ERROR_MAX_LEN, "Unknown error %x", msg);
-  }
-
-  if (kmcmp::ErrChr > 0) {
-    char *szTextNull = strchr(szText, 0);
-    snprintf(szTextNull, COMPILE_ERROR_MAX_LEN-(szTextNull-szText), " character offset: %d", kmcmp::ErrChr);
-  }
-
-  if (*ErrExtraLIB) {
-    char *szTextNull = strchr(szText, 0);
-    snprintf(szTextNull, COMPILE_ERROR_MAX_LEN-(szTextNull-szText), "%s", ErrExtraLIB);
-  }
-
-  ErrChr = 0;  *ErrExtraLIB =0;
-  if (!(*msgproc)(kmcmp::currentLine, msg, szText, msgprocContext)) return TRUE;
-  return FALSE;
-}
 
 KMX_DWORD ProcessBeginLine(PFILE_KEYBOARD fk, PKMX_WCHAR p)
 {
