+++ developer/src/kmcmplib/src/meson.build
@@ -72,6 +72,7 @@ lib = library('kmcmplib',
   'CheckNCapsConsistency.cpp',
   'CompileKeyboardBuffer.cpp',
   'Compiler.cpp',
+  'CompilerErrors.cpp',
   'CompilerInterfaces.cpp',
   'CompilerInterfacesWasm.cpp',
   'CompMsg.cpp',
